requirements.txt
appdirs==1.4.4
APScheduler==3.7.0
argon2-cffi==21.1.0
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
bleach==4.1.0
cachetools==4.2.4
causalgraphicalmodels==0.0.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
chromedriver-autoinstaller==0.2.2
click==8.0.1
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cryptography==35.0.0
cycler==0.10.0
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
dill==0.3.4
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.7.0
django-rest==0.8.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
entrypoints==0.3
filelock==3.0.12
google-api-core==2.0.1
google-api-python-client==2.23.0
google-auth==2.2.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
graphviz==0.18.2
gspread==4.0.1
gunicorn==20.1.0
h11==0.12.0
httplib2==0.19.1
idna==3.2
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.1
joblib==1.1.0
jsonfield==3.1.0
jsonschema==4.0.1
jupyter==1.0.0
jupyter-client==7.0.5
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
MechanicalSoup==1.1.0
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.4
numpy==1.21.2
oauth2client==4.1.3
oauthlib==3.1.1
otter-grader==3.1.4
outcome==1.1.0
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
patsy==0.5.2
pdfkit==0.6.1
pickleshare==0.7.5
Pillow==8.3.1
pipenv==2021.11.5.post0
platformdirs==2.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.0
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
Pygments==2.10.0
PyJWT==2.1.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-on-whales==0.27.0
pytz==2021.1
pywin32==301
pywinpty==1.1.4
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
schedule==1.1.0
scikit-learn==1.0.1
scipy==1.7.1
seaborn==0.11.2
selenium==4.0.0b4
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.1
statsmodels==0.13.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
toml==0.10.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
trio==0.19.0
trio-websocket==0.9.2
typer==0.4.0
typing-extensions==3.10.0.2
tzlocal==2.1
undetected-chromedriver==3.0.3
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.7.0
virtualenv-clone==0.5.7
wcwidth==0.2.5
webdriver-manager==3.4.2
webencodings==0.5.1
websockets==10.0
whitenoise==5.3.0
widgetsnbextension==3.5.1
wsproto==1.0.0
